lang_portal/backend_python/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
PyYAML>=6.0.1
uritemplate>=4.1.1
django-cors-headers>=4.3.1
invoke==2.2.0
python-dotenv==1.0.1
pydantic==2.6.1
sentry-sdk==1.39.1
setuptools>=69.0.3

# Development dependencies
black>=24.1.1
flake8>=7.0.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0

# Testing dependencies
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0 